Menambahkan utensil page, dan utensil contract

+++ lib/data/remote/models/ingredient_model.dart
@@ -5,7 +5,7 @@ part 'ingredient_model.g.dart';
 @JsonSerializable()
 class Ingredient {
   String? name;
-  int? quantity;
+  double? quantity;
   String? unit;
 
   Ingredient({
